Remove copying of unnecessary files to VSAC dir
I haven't tested this change but just checking out the work you guys have done and noticed this bash script. The files in https://github.com/siteadmin/code-validator-api/tree/master/codevalidator-api/docs with the exception of the subfolder ValueSetsHandCreatedbySITE are just informational for the user/not meant to be machine-readable/added to the VSAC folder. Only value sets should be in the VSAC folder as the validator will try to load them as such. This is why the one file previously had to be removed/caused in error if it was not. You could copy them to some other directory if that is useful to have local for a user or if you happen to be referencing these files elsewhere. By the way this is a really cool project. I will try it out soon and would like to further discuss it.